New stuff added.
Improved the BalancedTree and BalancedSet classes by implementing an
iterator. The iterator isn't the most memory efficient but, it does
detect modification of the underlying source. Since the iterator is
in-order, it iterates least to greatest.

Also updated JavaDoc

Also added size and height utility methods to the BalancedTree class.
Very useful when debugging.

Enjoy,

Vincent J. Palodichuk